requirements.txt
flask==2.3.2
python-dotenv==1.0.0
openai==1.35.0
requests==2.31.0
openpyxl==3.1.2
python-pptx==0.6.23
pandas==2.0.3
python-docx==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
llmwhisperer-client==0.2.0
matplotlib==3.7.2
plotly==5.15.0
seaborn==0.12.2
Pillow==10.0.0
numpy==1.24.3